Reorder tables
