+++ backend/apps/iamstudent/tables/student_list.py
@@ -2,7 +2,7 @@ from django.utils.html import format_html
 from django.utils.translation import gettext as _
 import django_tables2 as tables
 
-from .models import Student
+from apps.iamstudent.models import Student
 
 
 class StudentTable(tables.Table):
